Implement 'config migrate' command
